Fixes #30665 - drop the intl polyfills
drop the intl polyfills and rely on the  browser native intl
drop the old javascript webpack-vendor file as it only contain polyfills
