+++ app/assets/javascripts/application.js
@@ -69,12 +69,10 @@ function onContentLoad() {
 
   password_caps_lock_hint();
 
-  tfm.i18n.intl.ready.then(function() {
-    var tz = jstz.determine();
-    $.cookie('timezone', tz.name(), {
-      path: '/',
-      secure: location.protocol === 'https:',
-    });
+  var tz = jstz.determine();
+  $.cookie('timezone', tz.name(), {
+    path: '/',
+    secure: location.protocol === 'https:',
   });
 
   $('.full-value').SelectOnClick();
