--- config/webpack.vendor.js
@@ -1,1 +1,0 @@
-module.exports = ['react-intl', 'intl'];
