+++ webpack/assets/javascripts/react_app/common/I18n/I18n.test.js
@@ -1,7 +1,7 @@
 import Jed from 'jed';
-import { translate, ngettext } from './I18n';
+import { translate, ngettext } from './index';
 
-jest.unmock('./I18n');
+jest.unmock('./index');
 jest.unmock('jed');
 
 describe('gettext', () => {
