+++ webpack/assets/javascripts/react_app/common/I18n/index.js
@@ -1,0 +1,11 @@
+import intl from './intl';
+import { jed, translate, ngettext, sprintf } from './jed';
+
+export const documentLocale = () => intl.locale;
+
+export { translate, ngettext, jed, sprintf, intl };
+
+export default { translate, ngettext, jed, sprintf, intl };
+
+window.__ = translate;
+window.n__ = ngettext;
