+++ webpack/assets/javascripts/react_app/common/I18n/intl.js
@@ -1,0 +1,7 @@
+const htmlElement = document.getElementsByTagName('html')[0];
+
+export default {
+  locales: window.locales || {},
+  locale: htmlElement.getAttribute('lang') || 'en',
+  timezone: htmlElement.getAttribute('data-timezone') || 'UTC',
+};
