+++ webpack/assets/javascripts/react_app/common/i18nProviderWrapperFactory.js
@@ -9,32 +9,16 @@ const i18nProviderWrapperFactory = (
 ) => WrappedComponent => {
   const wrappedName = getDisplayName(WrappedComponent);
 
-  class I18nProviderWrapper extends React.Component {
-    constructor(props) {
-      super(props);
-      this.state = { i18nLoaded: false };
+  const I18nProviderWrapper = props => (
+    <IntlProvider
+      locale={intl.locale}
+      initialNow={initialNow}
+      timeZone={timezone || intl.timezone}
+    >
+      <WrappedComponent {...props} />
+    </IntlProvider>
+  );
 
-      // eslint-disable-next-line promise/prefer-await-to-then
-      intl.ready.then(() => {
-        this.setState({ i18nLoaded: true });
-      });
-    }
-
-    render() {
-      if (!this.state.i18nLoaded) {
-        return <span />;
-      }
-      return (
-        <IntlProvider
-          locale={intl.locale}
-          initialNow={initialNow}
-          timeZone={timezone || intl.timezone}
-        >
-          <WrappedComponent {...this.props} />
-        </IntlProvider>
-      );
-    }
-  }
   I18nProviderWrapper.displayName = `I18nProviderWrapper(${wrappedName})`;
 
   return I18nProviderWrapper;
