+++ webpack/assets/javascripts/react_app/components/common/dates/IsoDate.test.js
@@ -4,7 +4,6 @@ import { mount } from '@theforeman/test';
 import React from 'react';
 import IsoDate from './IsoDate';
 import { i18nProviderWrapperFactory } from '../../../common/i18nProviderWrapperFactory';
-import { intl } from '../../../common/I18n';
 
 describe('Date', () => {
   const date = new Date('2017-10-13 00:54:55 -1100');
@@ -16,10 +15,7 @@ describe('Date', () => {
       <IntlDate date={date} defaultValue="Default value" />
     );
 
-    intl.ready.then(() => {
-      wrapper.update();
-      expect(wrapper.find('IsoDate')).toMatchSnapshot();
-    });
+    expect(wrapper.find('IsoDate')).toMatchSnapshot();
   });
 
   it('renders default value', () => {
@@ -27,9 +23,6 @@ describe('Date', () => {
       <IntlDate date={null} defaultValue="Default value" />
     );
 
-    intl.ready.then(() => {
-      wrapper.update();
-      expect(wrapper.find('IsoDate')).toMatchSnapshot();
-    });
+    expect(wrapper.find('IsoDate')).toMatchSnapshot();
   });
 });
