+++ webpack/assets/javascripts/react_app/components/common/dates/ShortDateTime.test.js
@@ -4,7 +4,6 @@ import { mount } from '@theforeman/test';
 import React from 'react';
 import ShortDateTime from './ShortDateTime';
 import { i18nProviderWrapperFactory } from '../../../common/i18nProviderWrapperFactory';
-import { intl } from '../../../common/I18n';
 
 describe('ShortDateTime', () => {
   const date = new Date('2017-10-13 00:54:55 -1100');
@@ -16,10 +15,7 @@ describe('ShortDateTime', () => {
       <IntlDate date={date} defaultValue="Default value" />
     );
 
-    intl.ready.then(() => {
-      wrapper.update();
-      expect(wrapper.find('ShortDateTime')).toMatchSnapshot();
-    });
+    expect(wrapper.find('ShortDateTime')).toMatchSnapshot();
   });
 
   it('formats date with relative tooltip', () => {
@@ -31,10 +27,7 @@ describe('ShortDateTime', () => {
       />
     );
 
-    intl.ready.then(() => {
-      wrapper.update();
-      expect(wrapper.find('ShortDateTime')).toMatchSnapshot();
-    });
+    expect(wrapper.find('ShortDateTime')).toMatchSnapshot();
   });
 
   it('formats date with seconds', () => {
@@ -42,10 +35,7 @@ describe('ShortDateTime', () => {
       <IntlDate date={date} seconds defaultValue="Default value" />
     );
 
-    intl.ready.then(() => {
-      wrapper.update();
-      expect(wrapper.find('ShortDateTime')).toMatchSnapshot();
-    });
+    expect(wrapper.find('ShortDateTime')).toMatchSnapshot();
   });
 
   it('renders default value', () => {
@@ -53,9 +43,6 @@ describe('ShortDateTime', () => {
       <IntlDate date={null} defaultValue="Default value" />
     );
 
-    intl.ready.then(() => {
-      wrapper.update();
-      expect(wrapper.find('ShortDateTime')).toMatchSnapshot();
-    });
+    expect(wrapper.find('ShortDateTime')).toMatchSnapshot();
   });
 });
