+++ webpack/assets/javascripts/react_app/components/common/forms/DateTime/DateTime.js
@@ -3,9 +3,8 @@ import PropTypes from 'prop-types';
 import { FieldLevelHelp } from 'patternfly-react';
 import { Field } from 'formik';
 import DateTimePicker from '../../DateTimePicker/DateTimePicker';
-
 import CommonForm from '../CommonForm';
-import { documentLocale } from '../../../../common/I18n';
+import { intl } from '../../../../common/I18n';
 import './DateTimeOverrides.scss';
 
 const DateTime = ({
@@ -19,7 +18,7 @@ const DateTime = ({
   value,
   initialError,
 }) => {
-  const currentLocale = locale || documentLocale();
+  const currentLocale = locale || intl.locale;
 
   return (
     <Field
